DBZ-20 Recovering database history
