+++ debezium-connector-oracle/src/main/java/io/debezium/connector/oracle/OracleSchemaChangeEventEmitter.java
@@ -1,11 +1,14 @@
 package io.debezium.connector.oracle;
 
+import java.util.Set;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import io.debezium.pipeline.spi.SchemaChangeEventEmitter;
 import io.debezium.relational.Table;
 import io.debezium.relational.TableId;
+import io.debezium.relational.Tables;
 import io.debezium.schema.SchemaChangeEvent;
 import io.debezium.schema.SchemaChangeEvent.SchemaChangeEventType;
 import oracle.streams.DDLLCR;
@@ -27,10 +30,25 @@ public class OracleSchemaChangeEventEmitter implements SchemaChangeEventEmitter 
     @Override
     public void emitSchemaChangeEvent(Receiver receiver) throws InterruptedException {
         SchemaChangeEventType eventType = getSchemaChangeEventType();
-        if (eventType != null) {
-            Table table = new OracleDdlParser().parseCreateTable(tableId, ddlLcr.getDDLText());
-            receiver.schemaChangeEvent(new SchemaChangeEvent(offsetContext.getPartition(), offsetContext.getOffset(), ddlLcr.getSourceDatabaseName(), ddlLcr.getDDLText(), table, eventType, false));
+        if (eventType == null) {
+            return;
         }
+
+        Tables tables = new Tables();
+
+        OracleDdlParser parser = new OracleDdlParser();
+        parser.setCurrentDatabase(ddlLcr.getSourceDatabaseName());
+        parser.setCurrentSchema(ddlLcr.getObjectOwner());
+        parser.parse(ddlLcr.getDDLText(), tables);
+
+        Set<TableId> changedTableIds = tables.drainChanges();
+        if (changedTableIds.isEmpty()) {
+            throw new IllegalArgumentException("Couldn't parse DDL statement " + ddlLcr.getDDLText());
+        }
+
+        Table table = tables.forTable(tableId);
+
+        receiver.schemaChangeEvent(new SchemaChangeEvent(offsetContext.getPartition(), offsetContext.getOffset(), ddlLcr.getSourceDatabaseName(), ddlLcr.getObjectOwner(), ddlLcr.getDDLText(), table, eventType, false));
     }
 
     private SchemaChangeEventType getSchemaChangeEventType() {
