+++ debezium-core/src/main/java/io/debezium/relational/history/DatabaseHistory.java
@@ -99,7 +99,7 @@ public interface DatabaseHistory {
      */
     void record(Map<String, ?> source, Map<String, ?> position, String databaseName, String ddl) throws DatabaseHistoryException;
 
-    void record(Map<String, ?> source, Map<String, ?> position, String databaseName, String ddl, TableChanges changes) throws DatabaseHistoryException;
+    void record(Map<String, ?> source, Map<String, ?> position, String databaseName, String schemaName, String ddl, TableChanges changes) throws DatabaseHistoryException;
 
     /**
      * Recover the {@link Tables database schema} to a known point in its history. Note that it is possible to recover the
