+++ debezium-core/src/main/java/io/debezium/relational/history/HistoryRecord.java
@@ -8,6 +8,7 @@ package io.debezium.relational.history;
 import java.util.Map;
 import java.util.Map.Entry;
 
+import io.debezium.document.Array;
 import io.debezium.document.Document;
 
 public class HistoryRecord {
@@ -16,6 +17,7 @@ public class HistoryRecord {
         public static final String SOURCE = "source";
         public static final String POSITION = "position";
         public static final String DATABASE_NAME = "databaseName";
+        public static final String SCHEMA_NAME = "schemaName";
         public static final String DDL_STATEMENTS = "ddl";
         public static final String TABLE_CHANGES = "tableChanges";
     }
@@ -26,7 +28,7 @@ public class HistoryRecord {
         this.doc = document;
     }
 
-    public HistoryRecord(Map<String, ?> source, Map<String, ?> position, String databaseName, String ddl, TableChanges changes) {
+    public HistoryRecord(Map<String, ?> source, Map<String, ?> position, String databaseName, String schemaName, String ddl, TableChanges changes) {
         this.doc = Document.create();
 
         Document src = doc.setDocument(Fields.SOURCE);
@@ -46,6 +48,10 @@ public class HistoryRecord {
 
         if (databaseName != null) {
             doc.setString(Fields.DATABASE_NAME, databaseName);
+        }
+
+        if (schemaName != null) {
+            doc.setString(Fields.SCHEMA_NAME, schemaName);
         }
 
         if (ddl != null) {
@@ -74,8 +80,16 @@ public class HistoryRecord {
         return doc.getString(Fields.DATABASE_NAME);
     }
 
+    protected String schemaName() {
+        return doc.getString(Fields.SCHEMA_NAME);
+    }
+
     protected String ddl() {
         return doc.getString(Fields.DDL_STATEMENTS);
+    }
+
+    protected Array tableChanges() {
+        return doc.getArray(Fields.TABLE_CHANGES);
     }
 
     @Override
