+++ debezium-core/src/main/java/io/debezium/relational/history/TableChanges.java
@@ -6,17 +6,22 @@
 package io.debezium.relational.history;
 
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import io.debezium.document.Array;
+import io.debezium.document.Array.Entry;
 import io.debezium.document.Document;
 import io.debezium.document.Value;
 import io.debezium.relational.Column;
+import io.debezium.relational.ColumnEditor;
 import io.debezium.relational.Table;
+import io.debezium.relational.TableEditor;
 import io.debezium.relational.TableId;
+import io.debezium.relational.history.TableChanges.TableChange;
 
-public class TableChanges {
+public class TableChanges implements Iterable<TableChange> {
 
     private final List<TableChange> changes;
 
@@ -24,9 +29,28 @@ public class TableChanges {
         this.changes = new ArrayList<>();
     }
 
+    public static TableChanges fromArray(Array array) {
+        TableChanges tableChanges = new TableChanges();
+
+        for (Entry entry : array) {
+            TableChange change = TableChange.fromDocument(entry.getValue().asDocument());
+
+            if (change.getType() == TableChangeType.CREATE) {
+                tableChanges.create(change.table);
+            }
+        }
+
+        return tableChanges;
+    }
+
     public TableChanges create(Table table) {
         changes.add(new TableChange(TableChangeType.CREATE, table));
         return this;
+    }
+
+    @Override
+    public Iterator<TableChange> iterator() {
+        return changes.iterator();
     }
 
     public Array toArray() {
@@ -37,6 +61,33 @@ public class TableChanges {
 
         return Array.create(values);
     }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + changes.hashCode();
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        TableChanges other = (TableChanges) obj;
+
+        return changes.equals(other.changes);
+    }
+
+    @Override
+    public String toString() {
+        return "TableChanges [changes=" + changes + "]";
+    }
+
     public static class TableChange {
 
         private final TableChangeType type;
@@ -49,6 +100,30 @@ public class TableChanges {
             this.id = table.id();
         }
 
+        public static TableChange fromDocument(Document document) {
+            TableChangeType type = TableChangeType.valueOf(document.getString("type"));
+            TableId id = TableId.parse(document.getString("id"));
+            Table table = null;
+
+            if (type == TableChangeType.CREATE) {
+                table = fromDocument(id, document.getDocument("table"));
+            }
+
+            return new TableChange(type, table);
+        }
+
+        public TableChangeType getType() {
+            return type;
+        }
+
+        public TableId getId() {
+            return id;
+        }
+
+        public Table getTable() {
+            return table;
+        }
+
         public Document toDocument() {
             Document document = Document.create();
 
@@ -56,6 +131,42 @@ public class TableChanges {
             document.setString("id", id.toString());
             document.setDocument("table", toDocument(table));
             return document;
+        }
+
+        @Override
+        public int hashCode() {
+            final int prime = 31;
+            int result = 1;
+            result = prime * result + id.hashCode();
+            result = prime * result + ((table == null) ? 0 : table.hashCode());
+            result = prime * result + type.hashCode();
+            return result;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            if (this == obj)
+                return true;
+            if (obj == null)
+                return false;
+            if (getClass() != obj.getClass())
+                return false;
+            TableChange other = (TableChange) obj;
+            if (!id.equals(other.id))
+                return false;
+            if (table == null) {
+                if (other.table != null)
+                    return false;
+            } else if (!table.equals(other.table))
+                return false;
+            if (type != other.type)
+                return false;
+            return true;
+        }
+
+        @Override
+        public String toString() {
+            return "TableChange [type=" + type + ", id=" + id + ", table=" + table + "]";
         }
 
         private Document toDocument(Table table) {
@@ -102,6 +213,54 @@ public class TableChanges {
             document.setBoolean("generated", column.isGenerated());
 
             return document;
+        }
+
+        private static Table fromDocument(TableId id, Document document) {
+            TableEditor editor = Table.editor()
+                .tableId(id)
+                .setDefaultCharsetName(document.getString("defaultCharsetName"));
+
+            document.getArray("columns")
+                .streamValues()
+                .map(Value::asDocument)
+                .map(v -> {
+                    ColumnEditor columnEditor = Column.editor()
+                        .name(v.getString("name"))
+                        .jdbcType(v.getInteger("jdbcType"));
+
+                        Integer nativeType = v.getInteger("nativeType");
+                        if (nativeType != null) {
+                            columnEditor.nativeType(nativeType);
+                        }
+
+                       columnEditor.type(v.getString("typeName"), v.getString("typeExpression"))
+                           .charsetName(v.getString("charsetName"));
+
+                       Integer length = v.getInteger("length");
+                       if (length != null) {
+                           columnEditor.length(length);
+                       }
+
+                       Integer scale = v.getInteger("scale");
+                       if (scale != null) {
+                           columnEditor.scale(scale);
+                       }
+
+                       columnEditor.position(v.getInteger("position"))
+                           .optional(v.getBoolean("optional"))
+                           .autoIncremented(v.getBoolean("autoIncremented"))
+                           .generated(v.getBoolean("generated"));
+
+                    return columnEditor.create();
+                })
+                .forEach(editor::addColumn);
+
+            editor.setPrimaryKeyNames(document.getArray("primaryKeyColumnNames")
+                .streamValues()
+                .map(Value::asString)
+                .collect(Collectors.toList()));
+
+            return editor.create();
         }
     }
