+++ debezium-core/src/main/java/io/debezium/schema/SchemaChangeEvent.java
@@ -15,6 +15,7 @@ import io.debezium.relational.Table;
 public class SchemaChangeEvent {
 
     private final String database;
+    private final String schema;
     private final String ddl;
     private final Set<Table> tables;
     private final SchemaChangeEventType type;
@@ -22,14 +23,15 @@ public class SchemaChangeEvent {
     private final Map<String, ?> offset;
     private final boolean isFromSnapshot;
 
-    public SchemaChangeEvent(Map<String, ?> partition, Map<String, ?> offset, String database, String ddl, Table table, SchemaChangeEventType type, boolean isFromSnapshot) {
-        this(partition, offset, database, ddl, table != null ? Collections.singleton(table) : null, type, isFromSnapshot);
+    public SchemaChangeEvent(Map<String, ?> partition, Map<String, ?> offset, String database, String schema, String ddl, Table table, SchemaChangeEventType type, boolean isFromSnapshot) {
+        this(partition, offset, database, schema, ddl, table != null ? Collections.singleton(table) : null, type, isFromSnapshot);
     }
 
-    public SchemaChangeEvent(Map<String, ?> partition, Map<String, ?> offset, String database, String ddl, Set<Table> tables, SchemaChangeEventType type, boolean isFromSnapshot) {
+    public SchemaChangeEvent(Map<String, ?> partition, Map<String, ?> offset, String database, String schema, String ddl, Set<Table> tables, SchemaChangeEventType type, boolean isFromSnapshot) {
         this.partition = Objects.requireNonNull(partition, "partition must not be null");
         this.offset = Objects.requireNonNull(offset, "offset must not be null");
         this.database = Objects.requireNonNull(database, "database must not be null");
+        this.schema = Objects.requireNonNull(schema, "schema must not be null");
         this.ddl = Objects.requireNonNull(ddl, "ddl must not be null");
         this.tables = Objects.requireNonNull(tables, "tables must not be null");
         this.type = Objects.requireNonNull(type, "type must not be null");
@@ -46,6 +48,10 @@ public class SchemaChangeEvent {
 
     public String getDatabase() {
         return database;
+    }
+
+    public String getSchema() {
+        return schema;
     }
 
     public String getDdl() {
